backend/requirements.txt
Django>=4.2
djangorestframework
pymysql
pyjwt
cryptography
paho-mqtt

